requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
streamlit
requests
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart